backend/telegram-bot/requirements.txt
Flask==2.3.2
python-telegram-bot==20.5
requests==2.31.0
gunicorn==21.2.0
python-dotenv==1.0.0
